requirements.txt
biopython==1.83
pandas==2.2.1